+++ tests/__init__.py
@@ -30,6 +30,7 @@ SP_OPEN_PROC = "pyaud.utils.Subprocess._open_process"
 OS_GETCWD = "os.getcwd"
 README = Path("README.rst")
 PYAUD_PLUGINS_PLUGINS = "pyaud.plugins.plugins"
+TYPE_ERROR = "can only register one of the following:"
 
 
 class NoColorCapsys: